# src/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Swagger Library Manager - OpenAPI 3.0",
        "description": "Esta é a documentação da API para gerenciamento de uma biblioteca",
        "contact": {
            "email": "apiteam@swagger.io"
    },
        "version": "1.0.0"
},
    "servers": [
      {
        "url": "http://localhost:3333"
      }
    ],
...
    "paths": {
...
      "emprestimos/{id}":{
        "get": {
          "summary": "Obter total de empréstimos de um usuário",
          "description": "Obtém o total de empréstimos de um usuário",
          "operationId": "id",
          "parameters": [
            {
              "name": "user_id",
              "in": "query",
              "description": "ID do usuário para o qual o total de empréstimos está sendo obtido",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Total de empréstimos obtido com sucesso"
            }
          },
          "tags": ["emprestimos"]
        },
        "patch": {
          "summary": "Devolver livro emprestado",
          "description": "Atualiza o status de um livro emprestado para devolvido",
          "parameters": [
            {
              "name": "user_id",
              "in": "path",
              "description": "ID do usuário que está realizando a devolução",
              "required": true,
              "schema": {
                "type": "integer"
              }
            },
            {
              "name": "book_id",
              "in": "path",
              "description": "ID do livro que está sendo devolvido",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Livro devolvido com sucesso!"
            },
            "404": {
              "description": "Operação não relizada!"
            }
          },
          "tags": ["emprestimos"]
        }
      }
    },
...
}